done new modules for functions
