+++ js/components.js
@@ -1,0 +1,24 @@
+// В файле main.js напишите необходимые функции для создания массива из 25 сгенерированных объектов. Каждый объект массива — описание фотографии, опубликованной пользователем.
+export const description = [
+  'Описание 1',
+  'Описание 2',
+  'Описание 3',
+  'Описание 4'
+];
+
+export const message = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const GENERATE_NAME = [
+  'Mask',
+  'Alyona',
+  'Dima',
+  'Zhenya',
+  'Rodya'
+];
